# alem-ekub-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habte",
            "email": "habtadugna@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
...
        "/user-guarantee/{round-id}/{user-id}": {
            "post": {
                "tags": [
                    "userGuarantees"
                ],
                "operationId": "guaranteeUser",
                "parameters": [
                    {
                        "name": "round-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/user-guarantee/guarantors/{round-id}/{user-id}": {
            "get": {
                "tags": [
                    "userGuarantees"
                ],
                "operationId": "getGuarantors",
                "parameters": [
                    {
                        "name": "round-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-guarantee/guaranteed-users/{round-id}/{user-id}": {
            "get": {
                "tags": [
                    "userGuarantees"
                ],
                "operationId": "getGuaranteedUsers",
                "parameters": [
                    {
                        "name": "round-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-guarantee/allowed-guarantor/{ekub-id}/{version}": {
            "get": {
                "tags": [
                    "userGuarantees"
                ],
                "operationId": "isAllowedToBeGuarantor",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user-guarantee/cancel-guarantee/{round-id}/{guarantor-id}/{guaranteed-id}": {
            "delete": {
                "tags": [
                    "userGuarantees"
                ],
                "operationId": "cancelGuarantee",
                "parameters": [
                    {
                        "name": "round-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "guarantor-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "guaranteed-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "profilePicUrl": {
                        "type": "string"
                    },
                    "idCardImageUrl": {
                        "type": "string"
                    },
                    "guarantors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "ekubs": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ekubIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
            "BooleanResponse": {
                "type": "object",
                "properties": {
                    "result": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}